requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
mysqlclient==2.2.4
python-dotenv==1.0.1
quick-mailer==2022.2.22
sqlparse==0.5.0
typing_extensions==4.11.0
